ddc-mqtt/requirements.txt
pyyaml>=6.0
paho-mqtt>=1.6.1